refactor AR library by moving a number of variables from the surface to the graphics class
